# reenable the precompile generation for Distributed