--- a/contrib/generate_precompile.jl
+++ b/contrib/generate_precompile.jl
@@ -121,16 +121,13 @@ if Distributed !== nothing
     precompile(Tuple{typeof(Distributed.procs)})
     precompile(Tuple{typeof(Distributed.finalize_ref), Distributed.Future})
     """
-# This is disabled because it doesn't give much benefit
-# and the code in Distributed is poorly typed causing many invalidations
-#=
     precompile_script *= """
     using Distributed
     addprocs(2)
     pmap(x->iseven(x) ? 1 : 0, 1:4)
     @distributed (+) for i = 1:100 Int(rand(Bool)) end
+    @everywhere 1+1
     """
-=#
 end