This was disabled in https://github.com/JuliaLang/julia/pull/37816 with the reason that latency for Distributed should not be crucial and that there was a lot of badly typed code in Distributed leading to invalidations in other code when this was precompiled into the sysimage.

However, it is becoming apparent that this had a significantly larger effect than intended, see https://github.com/JuliaLang/julia/issues/39291 and https://discourse.julialang.org/t/why-does-julia-use-thousands-of-cpu-hours-to-compute-1-2/66041.

Therefore, I think it is best to reenable this.

Some timings:

Master:

```
julia -p 64 -e 'using Distributed; @everywhere 1+1'  285.89s user 9.65s system 604% cpu 48.914 total
```

PR:

```
./julia -p 64 -e 'using Distributed; @everywhere 1+1'  29.68s user 2.21s system 174% cpu 18.256 total
```

cc @moble, @vancleve, @algorithmx