@timholy 2021-09-08
It would be great if one of the heavy or affected users could take Distributed under a wing. Last I looked (and admittedly there have been a bunch of PRs since then...), Distributed had nice functionality but was a bit of a train wreck when it came to inferrability. Maybe it doesn't matter for performance, but maybe it does, and loading a few packages is pretty likely to invalidate big chunks of what gets precompiled here.

Still, if it solves real-world problems, no objections to putting this back. I just don't plan on fixing any of the invalidations myself :slightly_smiling_face:.