@GregPlowman 2021-09-12
I use 200 workers across 5 machines (20 local + 4 x 45 remote).
Adding the 200 workers takes over 3 minutes (might be different issue)
After workers are added, loading with `@everywhere using` takes another 7-8 minutes.