@GregPlowman 2021-09-13
> With or without this PR? If without, could you try with this one?

Sorry, I should have given more info.
Times were for Julia v1.6.2 release version.
Faster loading on Julia 1.7.0-beta4, see comparison below.
Unfortunately and somewhat embarrassingly, I don't know how to test using this PR.

|       | Julia 1.6.2<br>(seconds) | Julia 1.7.0-beta4<br>(seconds) |
| --- | ---------------------------:  | ----------------------------------: |
| Adding workers | 185 | 204 |
| Loading modules | 445 | 192 |
| Total | 630 | 396 |

Local machine running Windows 10. Remote machines are Window Server 2019.