@KristofferC 2021-09-15
>  Unfortunately and somewhat embarrassingly, I don't know how to test using this PR.

You can get it from: https://s3.amazonaws.com/julialangnightlies/assert_pretesting/winnt/x64/1.8/julia-12621148ff-win64.exe