@jebej 2021-09-15
Could we backport this to 1.7?