@GregPlowman 2021-09-16
> You can get it from: https://s3.amazonaws.com/julialangnightlies/assert_pretesting/winnt/x64/1.8/julia-12621148ff-win64.exe

Thanks Kristoffer.
I've updated table in previous post with times from this PR.
Adding workers is faster.
Loading modules is slower than 1.7.0-beta4.