@timholy 2022-03-12
There does not seem to be a lot of evidence that it helps. Sure, it's faster to add workers. But load just about any package and "poof," much of the code in Distributed gets invalidated and has to be recompiled before the next operation can succeed. (The "total" line above is essentially unaltered by this PR.)

Until someone sits down and fixes the inference problems in Distributed, this will only add noise for anyone hunting for invalidations but who may not be a big user of Distributed.