@vancleve 2022-03-13
Agreed with @timholy here about the benefit of this change.

@timholy, do you think fixing the inference problems will fix this problem though? Based on my experience detailed here, https://github.com/JuliaLang/julia/issues/39291#issuecomment-914648826, it seems like the compile time is scaling with the number of workers (CPUs on the same or different nodes).  

@moble suggests that its because julia is waiting for something or blocking after sending something to each worker (https://github.com/JuliaLang/julia/issues/39291#issuecomment-916172174); is that true and if so, is fixing that (making this process not blocking) actually the solution here?